Update of NetBeans Project
The original code has been modified but not removed. All orginal code
has been commented out to reflect what has changed vs. what was added.
The login and password verification is working. It validates connection
to the remote database. The mediator file has been modified to reflect
that it is connecting to the remote MYSQL database. Have added
libraries project that have helped with connectivity.